Move inertia share stuff into seperate provider.  Register services
