+++ app/Providers/InertiaServiceProvider.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Providers;
+
+use App\Presenters\UserPresenter;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
+use Illuminate\Support\ServiceProvider;
+use Inertia\Inertia;
+
+class InertiaServiceProvider extends ServiceProvider
+{
+    /**
+     * Register any application services.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        //
+    }
+
+    /**
+     * Bootstrap any application services.
+     *
+     * @return void
+     */
+    public function boot()
+    {
+        Inertia::share([
+            'errors' => function () {
+                return Session::get('errors')
+                    ? Session::get('errors')->getBag('default')->getMessages()
+                    : (object) [];
+            },
+            'auth' => function () {
+                return [
+                    'user' => UserPresenter::make(Auth::user())
+                        ->preset('summary')
+                        ->get()
+                ];
+            }
+        ]);
+    }
+}
